+++ test/src/cognitect/client/test_double_test.clj
@@ -69,4 +69,7 @@
   (testing "throws when instrumenting op unsupported by service"
     (is (thrown-with-msg? RuntimeException
                           #"Operation not supported"
-                          (client.test/client {:api :s3 :ops {:DoesNotExistInS3 {}}})))))
+                          (client.test/client {:api :s3 :ops {:DoesNotExistInS3 {}}}))))
+
+  (testing "supports limited keyword access"
+    (is (= "s3" (:api (client.test/client {:api :s3}))))))